+++ src/api/api.ts
@@ -237,9 +237,9 @@ export default class MastoApi {
     }
 
     // Collect and fully populate / dedup a collection of toots for an array of Tags
-    async getStatusesForTags(tags: MastodonTag[]): Promise<mastodon.v1.Status[]> {
+    async getStatusesForTags(tags: MastodonTag[], numTootsPerTag?: number): Promise<mastodon.v1.Status[]> {
         console.log(`[getStatusesForTags()] called for ${tags.length} tags:`, tags.map(t => t.name));
-        const tagToots = await Promise.all(tags.map(tag => this.getStatusesForTag(tag)));
+        const tagToots = await Promise.all(tags.map(tag => this.getStatusesForTag(tag, numTootsPerTag)));
         return tagToots.flat();
     }