+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # NEXT RELEASE
 
 ### v0.26.2
-* Don't set catchupCheckpoint if the feed is empty
+* Don't set `catchupCheckpoint` if the feed is empty
 * Fix bug where `Account.webfingerURI` field was not getting set
 
 ### v0.26.1